+++ CHANGELOG.rst
@@ -2,11 +2,18 @@
 Changelog
 =========
 
-GEOPHIRES v3 (2023-2025)
+GEOPHIRES v3 (2023-2026)
 ------------------------
+3.11
+^^^^
+
+3.11: `Fervo_Project_Cape-5: Multi-Year Construction & Reservoir Engineering updates; SAM-EM Payback Period; SAM-EM ITC; User-Provided Reservoir Profile updates; SAM-EM MCY Levelized Metrics update <https://github.com/NatLabRockies/GEOPHIRES-X/pull/457>`__ | `release <https://github.com/NREL/GEOPHIRES-X/releases/tag/v3.11.12>`__
+
 
 3.10
 ^^^^
+
+3.10.25: `Add Number of Injection Wells per Production Well parameter <https://github.com/softwareengineerprogrammer/GEOPHIRES/pull/119>`__
 
 3.10: `SAM Economic Models: Multiple Construction Years; Number of Fractures per Stimulated Well parameter; Royalty Rate Escalation Start Year parameter <https://github.com/NREL/GEOPHIRES-X/pull/440>`__ | `release <https://github.com/NREL/GEOPHIRES-X/releases/tag/v3.10.24>`__